% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocation.R
\name{getLocation}
\alias{getLocation}
\title{get genomic location by gene symbol}
\usage{
getLocation(symbol, txdb, org)
}
\arguments{
\item{symbol}{Gene symbol}

\item{txdb}{txdb will be used to extract the genes}

\item{org}{org package name}
}
\description{
given a gene name, get the genomic coordinates.
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(org.Hs.eg.db)
getLocation("HSPA8", TxDb.Hsapiens.UCSC.hg19.knownGene, "org.Hs.eg.db")
}
