% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIdeo.R
\name{gieStain}
\alias{gieStain}
\title{color sheme for the schema for Chromosome Band (Ideogram)}
\usage{
gieStain()
}
\value{
A character vector of colors
}
\description{
Describe the colors of giemsa stain results
}
\examples{
gieStain()

}
