% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOneIdeo.R
\name{gridPlot}
\alias{gridPlot}
\title{plot GRanges metadata}
\usage{
gridPlot(gr, gp, type, xscale)
}
\arguments{
\item{gr}{an object of \link[GenomicRanges]{GRanges} with metadata. All 
metadata must be numeric.}

\item{gp}{an object of \link[grid]{gpar}}

\item{type}{type of the figure, could be barplot, line, point and heatmap}

\item{xscale}{x scale of the viewport}
}
\description{
plot GRanges metadata for different types
}
