% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listChromosomes.R
\name{listChromosomes}
\alias{listChromosomes}
\title{List the available chromosome}
\usage{
listChromosomes(file, format = c("hic", "cool"))
}
\arguments{
\item{file}{character(1). File name of .hic or .cool/.mcool/.scool}

\item{format}{character(1). File format, "hic" or "cool".}
}
\description{
List the chromosomes available in the file.
}
\examples{
hicfile <- system.file("extdata", "test_chr22.hic", package="trackViewer")
listChromosomes(hicfile)
coolfile <- system.file("extdata", "test.mcool", package="trackViewer")
listChromosomes(coolfile, format="cool")
}
