% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listNormalizations.R
\name{listNormalizations}
\alias{listNormalizations}
\title{List the available normalizations}
\usage{
listNormalizations(file, format = c("hic", "cool"))
}
\arguments{
\item{file}{character(1). File name of .hic or .cool/.mcool/.scool}

\item{format}{character(1). File format, "hic" or "cool".}
}
\description{
List the normalizations available in the file.
}
\examples{
hicfile <- system.file("extdata", "test_chr22.hic", package="trackViewer")
listNormalizations(hicfile)
coolfile <- system.file("extdata", "test.mcool", package="trackViewer")
listNormalizations(coolfile, format="cool")
}
