% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeStyle.R
\name{optimizeStyle}
\alias{optimizeStyle}
\title{Optimize the style of plot}
\usage{
optimizeStyle(trackList, viewerStyle = trackViewerStyle(), theme = NULL)
}
\arguments{
\item{trackList}{An object of \code{\link{trackList}}}

\item{viewerStyle}{An object of \code{\link{trackViewerStyle}}}

\item{theme}{A character string. Could be "bw" or "col".}
}
\value{
a list of a \code{\link{trackList}} and a \code{\link{trackViewerStyle}}
}
\description{
Automatic optimize the stlye of trackViewer
}
\examples{
extdata <- system.file("extdata", package="trackViewer",
                       mustWork=TRUE)
files <- dir(extdata, ".wig")
tracks <- lapply(paste(extdata, files, sep="/"), 
                 importScore, format="WIG")
re <- optimizeStyle(trackList(tracks))
trackList <- re$tracks
viewerStyle <- re$style
}
\seealso{
See Also as \code{\link{viewTracks}}
}
