% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGInteractions.R
\name{plotGInteractions}
\alias{plotGInteractions}
\title{plot GInteractions}
\usage{
plotGInteractions(gi, range, feature.gr, ...)
}
\arguments{
\item{gi}{an object of \link[InteractionSet:GInteractions-class]{GInteractions}}

\item{range}{the region to plot. an object of \link[GenomicRanges:GRanges-class]{GRanges}}

\item{feature.gr}{the feature.gr to be added. an object of \link[GenomicRanges:GRanges-class]{GRanges}}

\item{...}{Not used.}
}
\description{
plot graph for GInteractions
}
\examples{
library(InteractionSet) 
gi <- readRDS(system.file("extdata", "gi.rds", package="trackViewer"))
range <- GRanges("chr2", IRanges(234500000, 235000000))
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
feature.gr <- genes(TxDb.Hsapiens.UCSC.hg19.knownGene)
feature.gr <- subsetByOverlaps(feature.gr, regions(gi))
feature.gr$col <- sample(1:7, length(feature.gr), replace=TRUE)
feature.gr$type <- sample(c("promoter", "enhancer", "gene"), 
                         length(feature.gr), replace=TRUE, 
                         prob=c(0.1, 0.2, 0.7))
plotGInteractions(gi, range, feature.gr)
}
