% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allclasses.R
\docType{class}
\name{pos-class}
\alias{pos-class}
\alias{pos}
\title{Class \code{"pos"}}
\description{
An object of class \code{"pos"} represents a point location
}
\section{Slots}{

\describe{
\item{\code{x}}{A \code{\link{numeric}} value, indicates the x position}

\item{\code{y}}{A \code{\link{numeric}} value, indicates the y position}

\item{\code{unit}}{\code{"character"} apecifying the units for the corresponding
numeric values. See \code{\link[grid]{unit}}}
}}

