% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readHicBpResolutions}
\alias{readHicBpResolutions}
\title{Function for reading basepair resolutions from .hic file}
\usage{
readHicBpResolutions(fname)
}
\arguments{
\item{fname}{path to .hic file}
}
\value{
Vector of basepair resolutions
}
\description{
Function for reading basepair resolutions from .hic file
}
\examples{
readHicBpResolutions(system.file("extdata", "test_chr22.hic", package = "trackViewer"))
}
