% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readHicChroms}
\alias{readHicChroms}
\title{Function for reading chromosomes from .hic file}
\usage{
readHicChroms(fname)
}
\arguments{
\item{fname}{path to .hic file}
}
\value{
Data frame of chromosome indices, names and lengths
}
\description{
Function for reading chromosomes from .hic file
}
\examples{
readHicChroms(system.file("extdata", "test_chr22.hic", package = "trackViewer"))
}
