% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view3dStructure.R
\name{view3dStructure}
\alias{view3dStructure}
\title{Plot xyz data in 2d or 3d}
\usage{
view3dStructure(
  p,
  k = 3,
  feature.gr,
  atacSig,
  lwd.backbone = 2,
  col.backbone = "gray",
  lwd.maxAtacSig = 8,
  reverseATACSig = TRUE,
  col.backbone_background = "gray70",
  alpha.backbone_background = 0.5,
  lwd.gene = 3,
  coor_mark_interval = 5e+05,
  col.coor = "black",
  show_coor = TRUE,
  coor_tick_unit = 50000,
  label_gene = TRUE,
  col.tension_line = "black",
  lwd.tension_line = 1,
  length.arrow = NULL,
  safe_text_force = 3,
  square = TRUE,
  ...
)
}
\arguments{
\item{p}{GRanges object with mcols x, y, or with z}

\item{k}{The dimension of plot. 2: 2d, 3: 3d.}

\item{feature.gr}{The annotation features to be added. An object of \link[GenomicRanges:GRanges-class]{GRanges}.}

\item{atacSig}{The ATAC-seq signals. An object of \link[GenomicRanges:GRanges-class]{GRanges} with scores or an object of \link{track}.}

\item{lwd.backbone, lwd.gene, lwd.tension_line, lwd.maxAtacSig}{Line width for the 
linker, gene, interaction node circle, the dashed line of interaction edges, the tension line and the maximal reversed ATAC signal.}

\item{col.backbone, col.backbone_background, col.tension_line, col.coor}{Color
for the DNA chain, the compact DNA chain, the node circle, the linker, the tension line and the coordinates marker.}

\item{reverseATACSig}{Plot the ATAC-seq signals in reverse values.}

\item{alpha.backbone_background}{Alpha channel for transparency of backbone background.}

\item{coor_mark_interval}{The coordinates marker interval. Numeric(1). Set to 0
to turn it off. The default value 1e5 means show coordinates every 0.1M bp.}

\item{show_coor}{Plot ticks in the line to show the DNA compact tension.}

\item{coor_tick_unit}{The bps for every ticks. Default is 1K.}

\item{label_gene}{Show gene symbol or not.}

\item{length.arrow}{Length of the edges of the arrow head (in inches).}

\item{safe_text_force}{The loops to avoid the text overlapping.}

\item{square}{A logical value that controls whether control points for the curve are created city-block fashion or obliquely. See \link[grid]{grid.curve}.}

\item{...}{Not used.}
}
\value{
Coordinates for 2d.
}
\description{
Plot xyz data with grid or rgl package.
}
\examples{
p <- readRDS(system.file('extdata', '4DNFI1UEG1HD.chr21.FLAMINGO.res.rds',
    package='trackViewer'))
feature.gr <- readRDS(system.file('extdata', '4DNFI1UEG1HD.feature.gr.rds',
    package='trackViewer'))
view3dStructure(p, k=3, feature.gr=feature.gr, length.arrow=unit(0.000006, 'native'))
}
