% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageGR.R
\name{coverageGR}
\alias{coverageGR}
\title{calculate coverage}
\usage{
coverageGR(gr)
}
\arguments{
\item{gr}{an object of RGanges, GAlignments or GAlignmentPairs}
}
\value{
an object of GRanges
}
\description{
calculate coverage for \code{\link[GenomicRanges]{GRanges}},
 \code{\link[GenomicAlignments]{GAlignments}} or 
 \code{\link[GenomicAlignments]{GAlignmentPairs}}
}
\examples{
bed <- system.file("extdata", "fox2.bed", package="trackViewer",
                   mustWork=TRUE)
fox2 <- importScore(bed)
fox2@dat <- coverageGR(fox2@dat)

}
\seealso{
See Also as \code{\link[IRanges]{coverage}}, 
\code{\link[GenomicRanges]{coverage-methods}}
}
