% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneTrack.R
\name{geneTrack}
\alias{geneTrack}
\title{track from TxDb}
\usage{
geneTrack(ids, txdb, type = c("gene", "transcript"))
}
\arguments{
\item{ids}{Gene IDs. A vector of character. It should be keys in txdb.}

\item{txdb}{An object of \code{\link[GenomicFeatures:TxDb-class]{TxDb}}}

\item{type}{Output type of track, "gene" or "transcript".}
}
\value{
An object of \link{track}
}
\description{
Generate a track object from TxDb by given gene ids
}
\examples{
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
geneTrack(c("3312", "3313"), TxDb.Hsapiens.UCSC.hg19.knownGene)

}
