% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importGInteractions.R
\name{importGInteractions}
\alias{importGInteractions}
\title{Reading data from a ginteractions, hic, cool, or validPairs file}
\usage{
importGInteractions(
  file,
  format = c("ginteractions", "hic", "cool", "validPairs"),
  ranges = GRanges(),
  ignore.strand = TRUE
)
}
\arguments{
\item{file}{The path to the file to read.}

\item{format}{The format of import file. Could be ginteractions, hic, cool or validPairs}

\item{ranges}{An object of \code{\link[GenomicRanges:GRanges-class]{GRanges}} to indicate
the range to be imported}

\item{ignore.strand}{ignore the strand or not when do filter. default TRUE}
}
\value{
a \code{\link{track}} object
}
\description{
Read a \code{\link{track}} object from a ginteractions, hic, cool, or validPairs file
}
\examples{
#import a ginteractions file
#gi <- system.file("extdata", "test.ginteractions.tsv", package="trackViewer",
#                       mustWork=TRUE)
#dat <- importGInteractions(file=gi, format="ginteractions",
#                   ranges=GRanges("chr7", IRanges(127471197, 127474697)))

##import a hic file
#hic <- system.file("extdata", "test.hic", package = "trackViewer",
#                       mustWork=TRUE)
#dat <- importGInteractions(file=hic, format="hic")

##import a cool file
#cool <- system.file("extdata", "test.cool", package = "trackViewer",
#                       mustWork=TRUE)
#dat <- importGInteractions(file=cool, format="cool")

##import a validPairs file
#validPairs <- system.file("extdata", "test.validPairs", package = "trackViewer",
#                       mustWork=TRUE)
#dat <- importGInteractions(file=validPairs, format="validPairs")

}
\seealso{
See Also as \code{\link{importBam}}, \code{\link{track}}, 
\code{\link{viewTracks}}
}
