% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGInteractions.R
\name{plotGInteractions}
\alias{plotGInteractions}
\title{plot GInteractions}
\usage{
plotGInteractions(
  gi,
  range,
  feature.gr,
  label_region = FALSE,
  lwd.backbone = 3,
  col.backbone = "cyan",
  col.backbone_background = "gray",
  lwd.gene = 2,
  lwd.nodeCircle = 1,
  col.nodeCircle = "#DDDDDD25",
  lwd.edge = 2,
  col.edge = "gray80",
  show_coor = 1e+05,
  col.coor = "black",
  show_tension = TRUE,
  tension_unit = 1000,
  label_gene = TRUE,
  col.tension_line = "black",
  ...
)
}
\arguments{
\item{gi}{An object of \link[InteractionSet:GInteractions-class]{GInteractions}}

\item{range}{The region to plot. an object of \link[GenomicRanges:GRanges-class]{GRanges}}

\item{feature.gr}{The annotation features to be added. An object of \link[GenomicRanges:GRanges-class]{GRanges}.}

\item{label_region}{Label the region node or not.}

\item{lwd.backbone, lwd.gene, lwd.nodeCircle, lwd.edge}{Line width for the 
linker, gene, interaction node circle and the dashed line of interaction edges.}

\item{col.backbone, col.backbone_background, col.nodeCircle, col.edge, col.tension_line, col.coor}{Color
for the DNA chain, the compact DNA chain, the node circle, the linker, the tension line and the coordinates marker.}

\item{show_coor}{Show the coordinates or not. Numeric(1). Set to 0
to turn it off. The default value 1e5 means show coordinates every 0.1M bp.}

\item{show_tension}{Plot ticks in the line to show the DNA compact tension.}

\item{tension_unit}{The bps for every ticks. Default is 1K.}

\item{label_gene}{Show gene symbol or not.}

\item{...}{Not used.}
}
\description{
plot graph for GInteractions
}
\examples{
library(InteractionSet) 
gi <- readRDS(system.file("extdata", "gi.rds", package="trackViewer"))
range <- GRanges("chr2", IRanges(234500000, 235000000))
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(org.Hs.eg.db)
feature.gr <- genes(TxDb.Hsapiens.UCSC.hg19.knownGene)
feature.gr <- subsetByOverlaps(feature.gr, regions(gi))
symbols <- mget(feature.gr$gene_id, org.Hs.egSYMBOL, ifnotfound=NA)
feature.gr$label <- vapply(symbols, function(.ele) .ele[1], character(1L))
feature.gr$col <- sample(1:7, length(feature.gr), replace=TRUE)
feature.gr$type <- sample(c("cRE", "gene"), 
                         length(feature.gr), replace=TRUE, 
                         prob=c(0.1, 0.9))
plotGInteractions(gi, range, feature.gr)
}
