% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGInteractions.R
\name{plotGInteractions}
\alias{plotGInteractions}
\title{plot GInteractions}
\usage{
plotGInteractions(gi, range, ...)
}
\arguments{
\item{gi}{an object of \link[InteractionSet:GInteractions-class]{GInteractions}}

\item{range}{the region to plot. an object of \link[GenomicRanges:GRanges-class]{GRanges}}

\item{...}{Not used.}
}
\description{
plot graph for GInteractions
}
\examples{
gi <- readRDS(system.file("extdata", "gi.rds", package="trackViewer"))
range <- GRanges("chr2", IRanges(1, 100000000))
plotGInteractions(gi, range)
}
